Remove cycles between Embed and STS instances
`Embed sub super` has the `STS super` super-class constraint, but `Embed sub
super` is almost always used as a super-class constraint of `Embed sub super`
itself. This is a cycle. My hypothesis is that these cycles are fine, *until*
you start relying on constraints that are implied by them, because the
corresponding dictionaries will be recursive too. I believe this is what caused
some tests to hang.

To remedy this, remove the `STS super` super-class constraint from the `Embed`
class. Next, remove the new redundant constraints and add the constraints that
are now missing.

Finally, remove the cycle between each pair of `Embed` and `STS` instances. For
example:

```diff
instance
  ( ..
    Embed (UTXO era) (UTXOW era),
    ..
  ) =>
  STS (UTXOW era)
  where
  ..

instance
  ( ..
-   STS (UTXOW era),
    ..
  ) =>
  Embed (UTXO era) (UTXOW era)
  where
  ..
```