TxBody for ShelleMA (Mary and Allegra) eras
* "Added the TxBody type with validity intervals and forge fields. Tied this
together with the Timelocks scripts. TxBody is newtype wrapped around a MemoBytes.
It exports a set of HasField instances appropriate for a TxBody. Also created a
test file with a minumum number of test, testing the HasField use and roundtripp
CBOR properties. Cleaned up and extended Jareds idea of using EraIndependentTxBody.
Extended this to all types with a HashAnnotated instance, which now adds an
associated type family HashIndex. Made all uses consistent with this approach,
Removed the function eraIndTxBodyHash, which can be replaced hashAnnotated."

* Remove dead code in comments

Co-authored-by: Nicholas Clarke <nicholas.clarke@iohk.io>