Introduce the PraosCrypto class
This class exists in consensus and bundles all *signable* constraints that the
Shelley(-based) ledger needs. Since these constraints all originate from the
ledger, it makes sense to just put them in the ledger.

Using this class (which implies `Crypto crypto`) also simplifies the API
slightly.

Why not make this a constraint synonym? Because with a class, one can easily
check whether a given crypto type satisfies all the required constraints by
providing an instance declaration:

```haskell
instance PraosCrypto C_Crypto
```

If one of the constraints required for `PraosCrypto` is not satisfied by
`C_Crypto`, then there will be an error at the instance declaration.

When using a constraint synonym, there is no instance to define. An error about
a missing constraint will be only be visible where `PraosCrypto crypto` is
instantiated to `C_Crypto`, e.g., somewhere in consensus (way too late) or in
the testsuite.