Add class each era must satisfy for consensus integration
This is *the* constraint that an era must satisfy in order to be used by
consensus.

As discussed in the previous commit, this is a class and not a constraint
synonym. This means that instances of it can (and should!) be defined in the
ledger, which requires all constraints to be satisfied at the site of the
instance definitions.

With a constraint synonym, the constraints are only checked when the synonym is
instantiated and used, which would be *in the consensus layer*. This would mean
that the ledger won't be alerted when the required constraints are accidentally
removed. Instead, the consensus layer would detect it when updating its
dependency on the ledger, which is *too late*. By having the instance in the
ledger, it is up to the ledger to make sure all required constraints keep being
satisfied.