Fill out translation Shelley->Allegra
There are two interesting cases:
- The WitnessSet. In this case we embed the current multi-script
witnesses into the appropriate constructor from the Allegra era.
- The TxBody. This translation is interesting, since it has to preserve
the serialisation. In general this is not possible. However, in the
Allegra/Mary eras we carefully construct the serialisation types such
that a TxBody which is backwards compatible with Shelley will be
serialised the same way. As such, we're allowed to perform this
translation.