fix ValueNotConservedUTxO serialization
The ValueNotConservedUTxO predicate failure contains two values of type
Coin (in the Shelley era, but Core.Value in general). Though the Coin
values in the ledger state are always within the proper bounds
(ie Word64), the failure could produce values outside of this range.
This caused a serialization error for ValueNotConservedUTxO, since we
do not allow the serialization of coins out of the Word64 range.

For the Coin type, we already had a wrapper DeltaCoin which can be
used when we want to intentionally serialize any Coin.
This is now generalized to a Torsor type class, which
ValueNotConservedUTxO now uses.

Resolves: CAD-2168