Fix thunk in Likelihood
`Likelihood` was using a `Seq` instead of a `StrictSeq`, which meant that its
elements weren't forced to WHNF. These `Likelihood`s are stored in the
`NonMyopic` data type, which is part of the ledger state, so there was a space
leak in the ledger state.

I found this space leak because serialising the mainnet ledger state to disk was
taking roughly 2 minutes on my machine! I noticed this when I rebuilt the ledger
snapshot from genesis and tried to shut down the node afterwards with CTRL-C,
which took way too long (it was busy writing the snapshot and thus forcing the
thunks).

With this fix, serialising the mainnet ledger state takes roughly 1 second.