Fix serialisation of Bimap
This PR addresses a number of issues related to Bimap serialisation:
- The existing serialisation format involved encoding both the forward
and reverse maps, leaving our on-disk state twice as large as it needed
to be.
- A change in #1886 resulted in the internal structure of Bimap
changing, and a consequent inadvertant change in the serialised format.
We want to avoid this wherever possible, since this entails all nodes
needing to rebuild the Ledger DB from scratch.

Consequently, we make the following changes:
- Newly serialised 'Bimap's will serialise only the forward map. This is
then decoded to populate both directions.
- The decoder will also accept the previously encoded format, except
that in this case it ignores the backwards encoding. This should allow
it to decode 'Bimap's encoded before #1886.

The current implementation of 'biMapFromList' does not look terribly
efficient; this commit does not (yet) address that.