two pool example
We add another example to the suite. In this example, there
are two pools, call them Alice and Bob's pools.
There is also a delegator named Carl. In this scenario,
both Alice and Bob's pool list Carl's reward account in the
pool certificate, though Carl is not an owner in either pool.
Additionally, Carl delegates to Alice's pool.