fix timelock upper bound evaluation
The evaluation of RequireTimeExpire had two problems:

- The comparison of the transaction time to live with the timelock
expiration should be less than or equal to, not less than.
What we want is that the transaction interval is contained in the
timelock interval, possibly equal to.  This is not to be confused with
the fact that the time to live marks the first slot that the
transaction is expired.
- The comparison had the transaction value and the timelock value
swapped.