"Make the GenEnv Era generic.
To make the GenEnv Era generic we introduce 2 new classes ScriptQuantify and ScriptClass.
These provide operations that will differ for each Era, and that matter when creating
and using a GenEnv.  These classes (along with instances for the ShelleyEra) can be
found in  Test.Shelley.Spec.Ledger.Generator.GenEra. Instances for the Mary and Allegra
Era are found in Test.Cardano.Ledger.EraBuffet. We propogated the class constraints
upto the top level tests that build TestTrees. This should work in conjunction with other
pull requests that make things other than the GenEnv Era generic."