Soft fork mechanism to referece major version
Soft forks need reference to the major version, so that, for instance,
version (3,0) does not revert changes made in version (2,1).

Originally we had thougth that we would introduce a notion of a
"starting major version", so that a testnet starting in the shelley era
would not have to start at version 2 0.  But this may lead to more
confusion, and now we propose just to hard code the version pair.