Adjust how we decide when to activate soft forks
We have changed our mind from saying that soft forks should be relative
to an era, to instead simply saying there is a known global mapping
between eras and protocol versions.

This makes a difference for the consensus mode that starts in the
Shelley era. We're not saying that has to start at version 2, rather
than at version 0. This allows us to say that Shelley is always protocol
version 2 rather than it depending on the consensus mode.