Rename "Goguen"->"Alonzo".
The term "Goguen" refers to a whole package of different functionality
which is being delivered over multiple eras. For example, multi-assets
are introduced in the "Mary" era. As such we remove the term "Goguen"
and use instead "Alonzo" to refer to the era introducing extended UTxO
via Plutus Core.

Alonzo Church was the inventor of the Lambda Calculus. It seems only
fitting that the era introducing Plutus core is thus named after him!