Parameterise era-independent types over the crypto instead of era
A bunch of types were parameterised over the era, while they were really
era-independent. It is confusing to parameterise something over the era while it
really is era-independent. Moreover, it requires much more needless
conversions/coercions.

The change that made it all possible is to parameterise `ScriptHash` by the
crypto instead of the era, using the `EraIndependentScript` tag. This meant
`EraIndependentWitVKey` was no longer needed.

By pushing this through, we get to the following (possibly incomplete) list of
all the types that were parameterised over the crypto instead of the era:

* Addr
* BlocksMade
* BootstrapAddress
* BootstrapWitness
* CompactAddr
* Credential
* DCert
* DPState
* DState
* DelegCert
* Delegation
* GenesisDelegCert
* InstantaneousRewards
* MIRCert
* MultiSig
* MultiSigRaw
* NonMyopic
* PState
* PaymentCredential
* PoolCert
* PoolParams
* RewardAccounts
* RewardAcnt
* RewardUpdate
* ScriptHash
* ShelleyGenesisStaking
* SnapShot
* SnapShots
* Stake
* StakeCredential
* StakeReference
* Timelock
* TimelockRaw
* TxId
* TxIn
* Value era
* Wdrl
* WitHashes
* WitVKey