Add a test for ApplyTxError round-tripping
This test revealed a problem with value serialisation, which wasn't
being caught by the existing value serialisation test because that test
only generated a very small range of values, rather than using the
serialisation generator. So additional drop that custom generator and
use the normal one. This shows up the value serialistion bug.

In addition, modify `property` to fail rather than error when the values
don't match. This allows (a) for shrinking, and (b) for showing the
result.

TODO:
- Modify 'propertyAnn' to additionally fail correctly.
- Modify this test harness to allow shrinking of values.