Fix Value deserialisation.
Sometimes there is the possibility of values lying outside our expected
range (due to a bug, for example). These show up most notably in error
messages, which must be serialised. The custom decoder which forces
things to be encoded as Int64 will barf on these; instead, we drop it
and use the standard decoder from `cborg`.

This fixes round-tripping of both Value and ApplyTxError.