Drop custom generator for Metadata.
Using custom generators in the roundtrip tests is a bad idea, since they
are not the generators which get composed into the higher-level tests
being done downstream (e.g. on some of the state types). For example, a
custom value serialisation generator for Values caused us to miss a bug
which then caused a test breakage in consensus, which was using the
`Arbitrary` generator.

We *do* use custom generators for our property tests, but in this case
they are always explicitly threaded, so we don't run the risk of missing
things.