formal spec errata: stability window discrepancy
There is a single use of the randomnessStabilisationWindow in both
the spec and the implementation, but they do not agree.
The spec uses it in the UPDN transition, the implementation uses it
in RUPD. We intended what was written in the spec.
There is, however, no problem with these values in Ouroboros Praos,
so the spec is changed to reflect what was implemented and a note is
made in the errata.