Use an EraMap to specify rules for the era
The way we implement the era-parametrised rules leads to a somewhat
annoying situation. We have two choices when creating an STS "instance".
We can either declare it's specific to a given era:

```
instance STS UTXO (ShelleyEra c)
```

Or we can declare it's valid for all eras:

```
instance STS UTXO era
```

Of course, what we'd really like is to say "this instance should be
valid for all eras going forth, until it's overridden". But this is a
little tricky to do at the type level, and we'd have to introduce some
dependently-typed wizardry to manage it.

This problem shows itself particularly with the update system, for
example. The update system is not modified in Allegra, Mary or Alonzo.
However, to support overriding it in a future era (which we need), we
would have to move from the "all era" style to the "current era only"
style. So the Shelley declaration would become valid for Shelley only,
and we'd have to introduce (identical) instances in Allegra, Mary and
Alonzo. This would both be substantial duplication and would muddy the
waters of what's actually overridden in each era.

This PR presents an alternative, "lightweight" solution to this problem,
based upon Damian's work in #2041. We add an additional layer of
indirection in the form of a type family which maps symbols (e.g.
"UTXO") into a specific implementation of the rule. We then have a map
of these per era (the EraMap). In future eras, we provide exactly this
same map, which identifies which rules come from which era. Obviously,
constraints guide which rules can be connected to each other.

This approach also removes the need for orphan STS instances in future
eras, since we would now define a "new" UTXO rule in a future era, for
example.