Remove the Torsor class
This purpose of this class was to enable type level distinctions between
values in TxOuts (which are always nonnegative) and values elsewhere
which may have either no constraints or other constraints.

We do not currently assign different types to these and don't have
plans to do so.

Instead there are runtime checks in the deserializers for the mint field
(no ada) and the txout field (no negative numbers).