used cached protocol parameters when aggregating
Hooray for property tests! They caught a bug!
In two places we were aggregating the rewards with the wrong protocol
parameters. Rewards are calculated useing a cached set of the parameters
so that if they change do to an update we are still using the correct
values. The cached values should be supplied to the aggregating function
instead of the current values.