Fix shelley-ma era tests.
For convenience's sake, I made some changes to `ShelleyTest`, adding the
`PParams ~ Core.PParams` constraint (which is needed almost everywhere,
since many things look at paramters or update them in the tests) and
dropping the body and multisig equality constraints, which are needed
rarely. But really this is all a giant mess and I'm very much looking
forward to our new test framework!