Fix performance regression in shelley-ma
PR #1943 addressed a problem that caused low performance when certain
types of transactions were rejected. However, this fix was only applied
in the Shelley era, not the MA era. Consequently, when we bumped to
Shelley MA, a regression was caused.

This commit applies the same fix as used in the Shelley era into the MA
eras.