Use Annotated decoder for PParams and PParamsDelta
This is necessary, since annotations will be required to decode these in
future eras, and we cannot/do not abstract over these kinds of
constraints.

One consequence of this is that we cannot use the `Trans_` constraints
for `FromCBOR`, since sometimes they require `FromCBOR` and sometimes
`FromCBOR (Annotator ..)`. Hence we have to unroll constraints in a
number of places. Some of the low-level (UTxO, Value) `Trans` constraint
synonyms are still in use, since these don't (currently) encompass
anything that requires annotation.