New MIR cert features in Alonzo
In the Alonzo era, MIR certs can be used to transfer funds between the
reserves and the treasury. Additionally, the semantics of the MIR certs
has changed when moving funds to stake accounts. Previously, if a reward
account was already in the _irwd mapping, it was overridden by new MIR
certs. Now, repetitions are handled by adding the values. Moreover,
negative values are allowed, so long as the change never results in a
negative value in the sum.

The type has changed for the MIR certs to accomodate the new changes,
and predicate failures use used to prevent the new features from being
used before the Alonzo era.