Added constraint free selectors in Alonzo era.
The Alonzo Era has a number of patterns that introduce selectors that
have unnecessary constraints. The GHC pattern for these types must serialize
arguments to construct a Memoized bytestring. Selectors decompose already constructed
objects and do not need these constraints, which the GHC pattern construct imposes.

For each pattern
pattern T :: Foo a => a -> b -> T
pattern T {a, b} <- TRaw a b where ...

we create an additional pattern, which primes the constructor and selectors.
pattern T' :: a -> b -> T
pattern T' {a', b'} <- TRaw a b

Thus selectors: a' and b' do not have the unnecceasary Foo constraint.
The module exports are also revised from:
T(T,a,b)   to   T(T,a,b,T',a',b').

By replacing many a with a', many functions have fewer constraints.
This commit handles all such cases in the Alonzo era code.

We had to lessen the constraints on the function
Shelley.Spec.Ledger.UTxO(balance) which was heavily over-constrained.
This required changing a few lines in only that file, since 'balance'
is now less-constrained it is still useable at all its old call sites.