Completed the Utxow rule, Completed Alonzo Scripts.
The Alonzo Script datatype had a stub for Plutus scripts. Imported the right type
and completed Show, Eq, PrettyA, ToCBOR and FromCBOR instances.

The Uxtow rule is now complete. Introduced AlonzoPredFail the UtxoW PredicateFailure type
for the Alonzo era. It embeds the PredicateFaliure type from the Shelley and ShelleyMa Eras.
Wrote Show, Eq, ToCBOR and FromCBOR instances. The Utxow rule now uses this type for Failures.

Generalized the ShelleyStyleWitness to lift from one PredicateFailure type to another so it
could be reused in any ERa that uses this style of PredicateFailure Embedding.