Remove an unevaluated thunk in RewardPulser
Spotted while looking through the reward pulsing code. The set of items
to pulse is stored as a list of tuples, and created by a call to
`Map.toList`. This list is then evaluated to WHNF by the strict bang in
`RewardPulser`, but this will only force the first list constructor.
Thus the thunks will remain for the majority of the reward computation.

This PR changes a couple of things to avoid this. Firstly, the
'RewardPulser' now stores its list of items to pulse as a `StrictSeq`.
Then we discard the tuple and store only the `PoolParams`, since the
other element of the tuple (the pool ID) is stored in the '`PoolParams`
anyway. Thus the bang in `RewardUpdate` should now be sufficient to
force evaluation to NF.

As part of this change, we generalise the `foldlM'` function to work
over any foldable structure. Note that since `toList` is lazy, the call
to `toList` should be O(1) via Data.Sequence.

Unfortunately Data.Coders had not been formatted and I accidentally
formatted it during this commit (I have format on save set). I'll
disentangle the formatting changes.

Note that this change will result in a change to the on-disk
serialisation format for `LedgerState`, and as such we should be careful
about when we merge it.