Added module Cardano.Ledger.Alonzo.FakePlutus
This module lets us experiment with small changes to the Plutus TxINfo and associated types.
We made a few changes to make the structure more closely align with the actual Ledger code.
Got close to completion on translating a Alonzo.Tx into a FakePlutus.TxInfo. Next step
mke the changes in FakePlutus real in Plutus.