Use new UtxoEnv type in Alonzo
In alonzo, the UTxOS rule needs to be able to resolve pointer address in
order to pass the stake credentials to the plutus interpreter.
Therefor the UtxoEnv type needs the pointer map.

This was nearly a small change, except that the UTXOW rule in all eras
uses the shelleyStyleWitness function, which was previous fixed to the
UtxoEnv type. The shelleyStyleWitness function is now parametric in this
type, but this came at the cost of applyBlock and applyTxs now needing
to this type to be specified.