Use wire encoding of redeemers in WitnessPPData
WitnessPPData is used in phase 1 validation to ensure that a script is
deterministic with respect to changes in protocol parameters. In
particular, it is hashed and checked against a hash in the transaction
body.

WitnessPPData consists of two main parts, the redeemers that are
included in the transaction and a language specific view of the protocol
parameters. When we construct the WitnessPPData, we preserve the
encoding of the redeemers from the original transaction, so that the
hash that we construct and compare with the hash in the body is correct.