Remove error calls, export type families from Era defining modules.
Filled in two calls to error in the Alonzo.hs file.
Made the Shelley, Allegra, Mary, and Alonzo files export type synonyms for
their type familes. So  Alonzo.TxBody  is the TxBody in the Alonzo era.
Added the file DescribeEras that adds descriptions of each Era. This file
won't compile if the descriptions are not accurate. Adds a much needed level
of documentation, about the reltinship between Eras.
Added a Ledger description of StandardCrypto in Cardano.Ledger.Crypto.