Introduce an era-generic fixed Tx type.
Per the discussions over the Alonzo transaction type, we came to the
conclusion that there is a key difference between transactions in the
mempool and transactions in the block. In general, downstream consumers
need care only about the _former_ concept. The node has to create them,
consensus has to manage them in the mempool.

Accordingly, rather than making `Tx` a type family, we choose to fix it
across eras, albeit with all of its component parts being era-indexed
families.

We would like to expose `Tx` from Core as with other things, so we make
use of an `hs-boot` file to allow the mutual recursion.