Check length plutus bytestring data, and double check some tests.
In order to validate data in both AuxiliaryData and in the WitnessSet, deserializing
a Plutus Data, with a ByteString type, whose length is greater than 64 now raises
a deserialisation error. Added a new combinator (<?) to Data.Coders that lets any
'constructor' raise an error. Also added roundtrip tests to be sure we actually
catch these cases.

Before running plutus scripts, we collect their input data using their hashes and the Witnesses.
Previous PredicateFailure tests should ensure we find Data for every script, BUT
the consequences of not finding Data means scripts can get dropped, so things
might validate that shouldn't. So we double check that every Script has its Data, and
if that is not the case, a PredicateFailure is raised in the Utxos rule.