Changed Alonzo Scripts to simple (no Memoized) sum type.
The Scrit in the Alonzo era has been changed to a simpe Sum type with two
constructors. This is possible because both variants store there original bytes,
so the SafeToHash instance does not need a Memobytes to memoize the originalBytes.

Also fixed up validting scripts, by using the Plutus side validateScript. this allowed
the repository to not need the serialize or the flat packages, which were removed
from the cabal file.