Alonzo Tx/TxInBlock separation.
This commit updates Alonzo to reflect the separation between Tx and
TxInBlock, and to implement 'SupportsSegWit'.

- Introduce a new Alonzo 'TxSeq' type along with the relevant
  serialisation, hashing, and functions to convert to a sequence of
  'ValidatedTx'.
- Modify existing code to make use of this.

Note that this commit requires dropping some of the API instances for
Alonzo; this is understandable, since those interfaces are now incorrect
(they do not reflect the difference between Tx and TxInBlock). A
subsequent PR will update the API to reflect the new view on the world.