Drop 'FromCBOR' for ValidatedTx.
The ValidatedTx type now exists only inside of blocks, and as such is
serialised only as part of a TxSeq. So we can drop the `FromCBOR`
instance entirely - it will never be used.

We _do_ still use the stored bytes, in order to compute the transaction
size. So for the moment we do still use the `ToCBOR` instances.