Benchmarks for 'aggregateUTxOCoinByCredential'
It was only after writing this that I noticed we already had some
benchmarks for this function, in

- https://github.com/input-output-hk/cardano-ledger-specs/blob/master/shelley/chain-and-ledger/shelley-spec-ledger-test/bench/BenchUTxOAggregate.hs
and
- https://github.com/input-output-hk/cardano-ledger-specs/blob/master/shelley/chain-and-ledger/shelley-spec-ledger-test/bench/Main.hs#L221

However, upon examining that benchmark, I think it's still worth having
this one (and maybe dropping the other). The existing benchmark uses
`gen` (I believe without fixing the seed), making it hard to test the
same case each time. And whilst we can control the size of the UTxO,
this new benchmark allows explicit control over the key things important
to the UTxO aggregation calculation, namely:
- The number of base stake addresses
- The number of pointer stake addresses (which require some additional
  computation), and
- The "delegation" factor; i.e. the number of TxOuts for each stake
  address.