Minor performance improvements in UTxO aggregation
This makes a couple of minor improvements in the function
'aggregateUtxoCoinByCredential', as measured by the new benchmark
introduced.

I tried a few variants; see
https://drive.google.com/drive/u/1/folders/1o7r4EF0uM5gfhtf3_GoMSJdoHc_d6Xtc
for the full reports. This version seems to shave maybe 20% off of the
20K UTxO run.

Interestingly, a variant where we try to explicitly deserialise only the
stake ref (using `deserialiseAddrStakeRef`) was _considerably_ slower.
I'm not immediately sure why!