Flip the 2-phase validation check in Utxos
Previously we checked whether scripts evaluated and then checked this
against the tag. This commit reverses that. This is only relevant
because the first check is done to determine rule dispatch, and thus
_must_ be evaluated. The second check is done to check transaction
validity. In our case, we would like to avoid performing those expensive
checks when re-applying a transaction, hence the switch.