Alonzo UTXOW rule should not call UTXO twice
The function shelleyStyleWitness calls the UTXO rule in every era.
Moreover, the Alonzo UTXOW rule calls both shelleyStyleWitness and the
UTXO rule directly (ignoring the state change in the first call).
This results in wasted computation and duplicate predicate failures.

We now remove the direct call of the UTXO rule in the alonzo UTOXW and
call shelleyStyleWitness at the end.