add a more coherent (less random) block generator
In 7447dd7b2c9af190f4c2d50fad83c42f68a41019 we changed genBlock to
produce blocks with more noise. This turns out to have broken some tests
in the consensus layer that relied on some coherency in the block.

See https://github.com/input-output-hk/ouroboros-network/issues/3134

To fix this, we add a new block generator genCoherentBlock, which is
almost exactly like the previous genBlock, but does not require the
PreAlonzo constraint.