Remove byte annotation from alonzo ValidatedTx
Preverving the bytes of this value is actually impossible.
There is no fact of the matter to preserve.
A user submits the following:
  [Body, Wits, Aux / Null]

We split this up and put the pieces into different placed in the block. We later
reassemble them for measuring transaction size. In this process the
original wrapper is lost, and we should not pretend it wasn't.

Also, this change brings the way we measure the size of the transaction
in line with how it was done in Mary. The ToCBOR instance of ValidatedTx
is now tested against Mary's.