Ensure we always have a collateral input.
In the case where we require collateral, we need to ensure that at least
one input gets spent, for replay protection purposes. Normally this will
be ensured by a collateral percentage > 0, but just in case, and to make
sure that we can't shoot ourselves in the foot, we check that if the Tx
_can_ fail in 2-phase validation, then we must have at least one
collateral input.