Avoid double-checking the 'IsValidating' flag.
Now that we run the newly constructed (via 'constructValidated')
'ValidatedTx' through the full LEDGER ruleset, we wish to avoid
re-evaluating scripts to check the tag which we have just created.

This commit achieves that by applying a special label to the 2-phase
script validation check, and disabling that check only when processing
the LEDGER rules in 'applyTx'.