Fix Plutus script encoding
The whole point of exposing a CBOR version was so we didn't expose the
internal format (which happens to be Flat right now). This was my
mistake since I did the wrong thing in the API module. I've fixed that
upstream and made sure that there's only one serialization typeclass
instance so we don't make this mistake again.

Nothing changes in the ledger proper since it already treated scripts as
opaque bytestrings, but some tests needed fixing.