Adjust the CBOR decoder for Rational to match CDDL
We use UnitInterval for several protocol params, but Rational for a_0
since it can be above 1.

For UnitInterval we use Ratio Word64 so that we stick to a bounded size
for the external representation.

For Rational we need to stick to a bounded external representation too,
and things are simpler if we use exactly the same one as for
UnitInterval. This is also the representation specified in the CDDL.

rational = #6.30([uint, uint])

This inconsistency was found by the generators and CBOR round trip tests
for the API package. It was using a Word64 ratio based generator for all
the Rational parameters. This failed for the a_0 param, since it only
accepts Int64 range. The a_0 param is not signed, so it does not need
signed integer support.

The name is now a bit of a misnoymer however. It might be better to make
a newtype wrapper for the a_0 param.