Add getUTxOSubset to the exported API
This will be used for a new node query to get a subset of the UTxO by
TxId, which makes it an efficient query (in both time and space).

Initially this is needed for the CLI to automate tx construction, but it
is general purpose and will likely find other uses.