Add getPools to the exported API and change getPoolParameters
The getPools returns the set of all pool ids.

The getPoolParameters is changed to work in a batch mode: taking a set
and returning a map, rather than looking up a single key. This change is
ok since there appear to be no existing users of getPoolParameters.

These will both be used for new node queries to get the set of all
currently registered pools, or pool params for selected pools.

Initially this is needed for the CLI to automate tx construction, but
they are general purpose and will likely find other uses.