cardano-ledger-core: parameterize the type of epochInfoWithError
When used with the HFC, Consensus currently stores this data type with a dummy
value in the `epochInfoWithError`. This has caused confusion, and current work
looks to be adding serialization, so we'd like to be able to use a more precise
type for the field (ie `()`) instead of a dummy value.