Drop 'TxInBlock' and use Core.Tx family.
Since `Tx` will now be what is submitted by the wallet/node, there is no
need for a separate `TxInBlock` concept. Instead, `Tx` becomes a type
family (since it _will_ vary between eras) and everything is defined in
those terms.

As part of this, we alter the `SupportsSegWit` type class to reflect now
the isomorphism between `TxSeq` and `Seq Tx`.