Rewrite mempool API in light of Tx/TxInBlock.
Previously, we used the Tx/TxInBlock distinction to also embody the fact
that a Tx had been validated, and hence future applications of that Tx
could skip a certain number of checks, an important performance concern.

Now we lack this distinction (which in any case was somewhat a
conflation of concerns), so we introduce the `Validated` newtype to
reflect this, and rewrite the Mempool API to reflect this.