WIP Trying to remove `Core.Value era ~ Value (Crypto era)` constraints.
This is important as we want to support any `Core.Value era` type.
Unfortunately, This seems like a show stopper in terms of getting to use
the CBOR representation of multi-asset Value. We'd need a From/ToCBOR
constraint on the value type, but this will probably be a bit of a
headache at the use sites, and also seems morally wrong.