Plutus to fail phase 1 validation on Byron addr
If a byron address is present in a transaction, instead of being
silently ignored (as is done in plutus V1), it now causes a phase 1 validation
error.  There are also now phase 1 validation errors for the logic
errors in translation that we do not expect to ever happen.

This change will happen at the next hard fork, major protocol version 7,
for both plutus V1 and V2.