computeStabilityWindow should not use to Double
There is no good reason to convert to a double before taking the ceiling
in this function. We now use Rational instead. Even though this is
actually a different function now, we can get away with this since both
mainnet and testnet have only ever had a single value which happens to
agree in both versions. In particular, mainnet and testnet have always
had k = 2160 and f = 1/20.

Similarly for computeRandomnessStabilisationWindow.