Set `TranslationContext` for Babbage to AlonzoGenesis.
This is a bit wrong, but makes downstream integration much easier. The
translation context is a diff from the previous era, whereas the genesis
is really the delta from Shelley. However, consensus currently assumes
that these are the same thing, since they have been for all prior eras.

At some point, we should alter consensus to drop this assumption and
deal correctly with the link between these. For now, we just set Babbage
to use the Alonzo genesis for its translation context.